Remove SMTP logging handler
This RabbitMQ consumer runs on many nodes. When the RabbitMQ cluster is unreachable, all RabbitMQ consumers send mail. This causes too much load on the mail platform. We considered using Bugsnag to keep track of errors, but we would have to specify the API key, which would then be visible by end users as they have root access. Taking into account that any errors in response processing are already kept track of by the caller, the Cluster API, we decided to remove the SMTP logging handler.