Make config access stable
As we will access the config from another package (for documentation generation), it must be stable. An undocumented dict is not, so switch to dataclasses.

Second, this commit fixes a side effect. The communication with RabbitMQ was intertwined with the config class (`RabbitMQ`). When getting the config, this caused unnecessary and unwanted communication with RabbitMQ. These concerns are now split into the `RabbitMQ` and `Config` classes, allowing for config access without side effects.

As tests will be rewritten properly in the future, and in the intermediate are completely broken as a result of this change, they have been removed.

Third, this commit adds `fernet_key` to the in-tree configs, serving as an example.

Finally, the environment variable to specify the config file path has been dropped. It has been replaced by a stable CLI argument.